Revise "Unified Collections" proposal.
This responds to all of the small-scale feedback on the pull request.
There are still a few things here and there I need to take into account.

The bigger question about what we want spread to desugar too is still an
open question that I want to explore more, but I wanted to get this out
first.