Updated according to discussion.
Extension names may conflict without affecting the implicit use of their declared members. Only access to the *name* is an error (and that can be double-imported with a prefix if necessary).

Pre-NNBBD, an `on` type of `Object` or `Null` will allow a `null` value at run-time, even without a trailing `?`.

Removed the "explicit extension" variants. We are not going with those.