Add new specification for `.packages` file
Provides a single source of truth about the `.packages` file format.

Updates the existing spec by removing references to `packages` directories (dropped in Dart 2) and adding metadata support for language versioning. Also changes the wording from being a proposal with motivation and arguments for making a particular choice, to just a specification of a specific design.