Transfer the string rules from Dart.g to the language specification (#629)
The string literal rules in the language specification relied on a number of undefined features that are not context free (like negation of multi-char tokens), and it did nothing to ensure proper nesting of non-simple string interpolations (in `${e}`, `e` can contain a string literal, which can contain a string interpolation, etc). This PR updates the rules to fix these issues, in particular by mentioning that we need a separate stack outside the parser/lexer in order to track nesting of interpolations.

This PR also transfers several grammar rules from Dart.g in order to eliminate an ambiguity in `<constObjectExpression>`, cf. issue 22393.