Specify that a constant e.length cannot use an extension method (#643)
Specify that a constant `e.length` cannot call an extension getter, and similarly for all the operator expressions. Also adjusted the treatment of negation and preincrement/decrement, to eliminate a syntactic issue, and in order to ensure that we specify both the static analysis and the dynamic semantics for these expressions.