Zeroth draft of design for patterns.
This needs a *lot* of work. But I'm at the point where I could use some
design feedback and help to figure out where to go next. The main
interesting things in here, I think are:

1. How it addresses bare identifiers.

2. How it addresses downcasts.

3. The protocol that tuple and custom extractor patterns use to pull
   data out of objects.

Those are what I'd like feedback on most, but of course all thoughts
are welcome.