Clarify the notion of being noSuchMethod forwarded (#817)
The existing definition conflated a couple of concepts (like using the same symbol for a name and a member and a declaration). The changes improve precision and eliminate some sources of confusion and misunderstandings, but the intended rules are unchanged (for instance, the example code illustrating a relevant situation is also unchanged).