Define T bounded types, to allow f() where f has type X extends Function (#900)
Define T bounded types and use that to allow f() where f has type X extends Function, and such. Also add a rule that a `C` bounded type has the interface of `C`, which is needed in order to allow `x.foo()` where `x` has type `X extends C` and `foo` is a member in the interface of `C`. We never said that previously, but it has been the implemented behavior as long as we have had the relevant types (type variables, and later on promoted types).