Adjust mixed mode rules (#897)
Adjust text about mixed-mode inheritance, to ensure that legacy classes have associated properties (class interface and implemented interfaces) that are purely legacy. Added explicit statement that the canonical interface implemented by a class with conflicts in the superinterface graph (like implementing both `A<int>` and `A<int?>`) is used both for dynamic instance checks and for static subtype checks.