Update to new number operation rules (#1065)
Rewrote rules to better preserve the type of the first operand. This allows `x+=y;` to work as much as possible, no matter which subtype of `num` the static type of `x` is (always when`y` is an integer or if `x` is a double. Use LUB for the type of `clamp`.
The rules are also simpler now.