Adjust UP for null safety, cf. language issue #1193 (#1195)
As mentioned in #1193, the rules for UP involving type variables need to be adjusted for soundness. This PR makes an adjustment to achieve this, using the greatest closure when the bound is an F-bound. It also adjusts the rules for UP(T Function<...>(...), T2) and UP(T1, T Function<...>(...)), where the current rule rule may incorrectly discard nullability.