Specify that null shorting does not include operators (#1256)
Specify that the null shorting transformation does not consider operator invocations as member invocations, cf. #1081. This implies that a construct like `-a?.b` is an error because `a?.b` can be null and `-` is not defined on a nullable receiver. If, following the language specification, we consider `-a?.b` to be syntactic sugar for `a?.b.unaryMinus()`, where `unaryMinus` is an instance method with the same behavior as `unary-`, we get a different behavior: `-a?.b` would then be allowed, and `-` / `unaryMinus()` would be skipped if `a` is null, so we need to say that this equivalence does not apply for null shorting. The same situation arises for other operators, e.g., additive and multiplicative ones.