Revise spec section imports (rebased PR) (#1161)
Section `Imports` and many concepts in there were in need of a substantial update (e.g., the treatment of prefixed namespaces was clearly not correct).

This PR introduces a new section `The Imported Namespace` which specifies how to handle conflicts (imported names with no prefix, or with a prefix, or conflicts involving a prefix and a local declaration). It contains a less substantial revision of the material about the semantics of imports, now in a separate section. It adds a section on 'Namespace Combinators' (show and hide), and a section about 'Conflict Merging of Namespaces', handling name clashes in a uniform manner where needed.

There is no implementation effort based on this PR: The updates amount to a specification and clarification of existing behavior. There is one exception (conflict resolution involving declarations with the same basename), but that area is handled separately, with tests, implementation issues, and breaking change management handled separately.