Adjust grammar to allow metadata of the form `@variableName` (#1344)
Cf. language issue #1341: The language specification grammar did not allow for metadata of the form `@<identifier>`, it only allowed `@<typeIdentifier>`; for example, this makes `@abstract` a syntax error even though it is possible to have a constant variable named `abstract`. This PR corrects the grammar to allow `@<identifier>`. The syntax of other forms of metadata (e.g., `@MyClass(42)`) is unchanged.