It is an error to create an instance of a type alias that expands to a type variable (#1361)
This PR adds the rule to the feature specification of non-function type aliases that it is an error to have an instance creation or a redirecting factory constructor body which is a type alias that expands to a type variable.

That's a new concept, e.g.: with `typedef F<X> = X;` and `typedef G<X, Y> = F<Y>;`, both `F` and `G` expand to a type variable.

The motivation for this error is that a type argument is a type, and it is not necessarily associated with a class, and instance creation is only supported for classes. (For example, the value of a type variable `X` could be the value of another type variable `Y`, and we don't support `new Y()` which would be the meaning of `new F<Y>()`).