Eliminate the static scope of a class (#1400)
This PR eliminates the static and instance scopes of a class and adds a new scope which replaces both of them (the 'class' scope).

It also adjusts the terminology to say 'current scope' rather than 'enclosing scope' in a number of situations where the target is a term rather than a scope (an expression has a 'current' scope, only a scope has an 'enclosing' scope). We still talk about 'the enclosing scopes' of a term or expression $t$, meaning the current scope $S$ of $t$ as well as the enclosing scopes of $S$, to keep down the verbosity.

The PR also defines the treatment of constructors in scopes: They are never added to any scopes. The class is looked up using the usual scoping mechanism, but the constructor is then directly accessed as a declaration in the given class, not by looking up anything in a scope. (It has been unclear since 2013 how constructors are treated by scopes, finally it's done! ;-).