Clean up \commentary, \rationale, specify \DefineSymbol (#1485)
Correct all \commentary{} and \rationale{} to have no space at front/end (to avoid spurious empty lines). Add `\DefineSymbol{}` in sections until and including 'Enums', in order to support easy scanning of the text to find the defining occurrence of any symbol which is used over a span of more than a few lines of text.