Add/clarify type alias errors in specification (#1468)
This PR updates the type alias section in the language specification to include an explicit rule about how to determine that there is a cyclicity error ("`F` depends on itself").

It also adds a specification of the error where a type alias is used in an instance creation, but it expands to a type variable. (Like typedef `F<X> = X;`), and adds a new rule that instance creations using a type alias must be regular-bounded (so with `class C<X> {}` and `typedef F<X extends num> = C<X>;`, `C<dynamic>()` and A<int>()` are (of course) OK, but `A<dynamic>()` is an error).