Adjust lexical lookup to include implicit extension method invocation (#1501)
The section on lexical lookup was rewritten such that a lexical lookup will now (1) encounter a compile-time error, or (2) yield a result, which can be (2a) a declaration, (2b) an import prefix, or (2c) nothing. The "caller" does not deal with situation (1) (we never specify propagation of errors), but it uses the result in situation (2a) and (2b), and it prepends `this` in situation (2c).

This is a bit different from previous approaches, because we transform the code and then run the static analysis on the transformed code, but in the case where we prepend `this.` it is actually never a different kind of error.

As a result, the change to lexical lookup is actually _not_ to specify how to handle extension method invocations in the section 'Lexical Lookup', it is to _remove_ the elements from that section specifying how to deal with instance method invocations. As a result, this part of the specification is now noticeably simpler and clearer.