Correct `part of` syntax (#1615)
Tiny bug fix, cf. dart-lang/sdk#45916:

The specification currently allows for a part of directive where the enclosing library is specified using <identifier> ('.' <identifier>)*, but not using <uri>. The latter is intended to work, and it has been working in all tools for a while, and this PR just brings the specification up to date.