Add paragraph to make it explicit that "let $S$ be a class" includes the case where it goes via a type alias (#1644)
Before this commit, the section 'Type Aliases' of the language specification specifies that it is an error if a type alias expands to a type variable, and it is used as a mixin. So type aliases expanding to a mixin are already taken into consideration.

However, it is not specified at various usage locations that type aliases are included. They are mentioned in connection with `extends` and `implements` and `with`, but, for instance, not in the section about mixin application.

This commit adds a paragraph where it is stated that it is in general implied that a term `$S$` can be a type alias that transitively expands to a class when we say things like "let `$S$` be a class".