Mention cascades in constructor-tearoffs feature spec (#1746)
Cascades were previously not mentioned in the constructor-tearoffs feature spec. This PR adds a paragraph about this topic: Constructs like `x..m<int>..m<String>.something` are allowed. An example is added to illustrate that `something` could be an extension member (so it's really marginal, and probably not very useful, but it isn't _just_ the case where `something` is a member of `Object`).