Macros: add Resource class specification for reading external resources (#1820)
Here is my WIP specification for how external resources can be loaded.

Note that it is vague enough to allow for reading http and potentially custom multi-root resources. This adds some general flexibility for the future, for instance if we wanted to run a compiler in a browser and load sources through http uris. In that case loading resources from http uris that fall under one of those package roots would be allowed.

I am a bit concerned about symlinks in general here still, I would love to hear thoughts on that.

Note that this also means macros can't load resources if there is no package config at all. So standalone dart scripts wouldn't be able to use macros that load resources, unless we carved out some sort of a special exception for this case.