Allow generic function instantiation of an existing function object (#1812)
Update section "Generic Function Instantiation" to allow this operation on an existing function object, in addition to the currently supported cases (where the expression denotes a declaration). Identity and equality is now specified compositionally (same function object and same type arguments implies same instantiated function object).

Remove the rule that it is an error to do generic method instantiation on the `call` of a function object (that error was introduced because we expected that it would be costly to perform generic instantiation on an existing function object; but this PR is all about _supporting_ that feature, because it turned out to be an unnecessary restriction).

Note that this update does not yet mention explicit instantiation, that is, it does not include any part of the feature bundle associated with the experiment 'constructor-tearoffs'.