Specify that C.m, C.m() cannot invoke an extension method (#1838)
It has been an error for years to use `int.toString()` to call the `Object` instance method toString on an instance of `Type`. This syntax ("`<typeLiteral> '.' <memberAccess>`") is reserved for static member access.

It is inconsistent if we allow the same syntax to invoke an extension method (because an extension method will always be dominated by an instance method on the static type of the receiver, and hence extension methods are presumed to always be invoked in a situation where we could also invoke an instance method).

The current specification wording does not make this obvious, and you could argue that it allows for `int.myExtensionMethod()` to invoke an extension method on `Type`.

This PR makes it explicit and unambiguous that it is not possible to implicitly call an extension method on a type literal.