Specify callable object closurization (#1842)
This PR adds a new case to the set of possible 'property extraction' expressions: Implicit property extraction. It occurs where an expression e has a static type as a callable object, and the context type is a function type. In this case we desugar e into e.call. Note that this will also serve as the first step in a generic function instantiation, because it is still true that the context type is a function type also in the case where e.call isn't assignable to that context type (but we can fix it by means of an implicit instantiation).

Smaller corrections:

The subsubsection about generic function instantiation was turned into a subsection: It used to be the case that a generic function instantiation would necessarily include a function closurization as a sub-action. This is no longer true, we can perform the generic instantiation on an existing function object. This means that the generic instantiation material should no longer be a subsection of 'Function Closurization', it's a stand-alone feature.

Subtitle uses version 2.15-dev, not 2.13-dev.

Renamed 'ordinary member closurization' to 'instance method closurization'. Given that we do not have getter/setter/operator tearoffs, instance member closurization only occurs for instance methods, and closurization of static members is handled under in the section 'Function Closurization' (and closurization of constructors will get their own section, too). So it's simply confusing and misleading to call it 'ordinary member closurization'.