Create proposal for enhanced enums.
Put this under issue 158 (Allow enums to store values), but it covers a large number of independent requests against enums.

Enums are already a shorthand for creating a class with a finite number of instances. This proposal makes it actually match classes for what it can declare of members. 
It does not allow individual values to declare members of their own, or override the default member (Java's enums does that). It's something which can be added in the future without risking backwards compatibility.