Reorder the sections in the macros proposal.
Because the diff of this is a nightmare, I have made no substantial
changes. All I did was reorganize the existing sections and add some
TODOs for parts to fill in. This means the resulting prose doesn't
flow and is weird in some places. I'll fix that in follow up PRs, but
I wanted to do the reordering separately since the diff is unreadable.

I also removed the TOC from the top since GitHub generates one
automatically: https://github.com/isaacs/github/issues/215#issuecomment-808199023