Edit "Generating Code" in the macros proposal. (#1909)
* Edit "Generating Code" in the macros proposal.

The diff is nasty because I reordered some sections (sorry). There are
no substantive changes to the proposal itself. Mostly just filling in
some more and trying to get it to flow a little better.

* Update intro to "generating macro applications".

The current design does require macros that declare and implement to
apply themselves in the declaration phase, so get the prose in line
with that and leave a TODO to consider changing that approach.