Alternate macro api proposal, less visible phases (#1964)
In this proposal, there are no explicit classes you implement for each phase, there is only a single class to implement based on the type of declaration you will annotate.

Each of these classes has a single visit method which gets a declaration argument and a context argument. The context has methods like `buildTypes`, `buildDeclarations`, and `buildDefinition`. These take a callback as an argument, and the callback receives the corresponding builder. It won't be executed until in the right phase.

Some advantages of this api are:

- Can have some shared state between phases, managed by local variables in the `visit` method.
- Less complicated class  heirarchy - far fewer classes to understand.
- The phases are less directly exposed, you don't really have to understand them so much any more.

Some disadvantages of this api are:

- Can result in runtime errors if violating phases - you can't call the `buildTypes` method from within the callback passed to `buildDeclarations` for instance. That would violate ordering, and we would have to throw at runtime. We could lint for this though.
- People may try to share state that isn't shareable across phases (for instance store the TypeBuilder and use it in the declaration phase).
- If only implementing one phase there is more boilerplate.