Update enhanced enum feature specification.
State that `index` and `toString` are inherited from `Enum`.
State that `values` is only added if it would not conflict.

Rephrase the definition in terms of an `enum` declaration
introducing a *class*, not a `class` declaration.
Then put the corresponding class declaration into the
"implementation" section as a desugaring, not a specification.