[patterns] Change named field shorthand syntax.
Instead of omitting the variable and inferring it from the field name,
we omit the field name and infer it from the variable. This allows all
types of variable patterns (var, final, typed), allows them to be
wrapped in null-check and null-assert patterns, and also allows
inferring a field name from a cast pattern.

Fix #2231.