Breaking change to await, update to flatten.
Changes **flatten** to be specified for a static type which is a type variable.

Changes `await` to throw when provided a `Future` value
of a type that could provide an unsound value if awaited.

This differes from the current specification, which instead treats the future as
a value, so `await someFuture` can evaluate to `someFuture`.

It also differs from the current implementation,
which awaits `someFuture` to its value,
whether the result is sound or not.

The result *can* be sound. Compare:
```dart
Object o = await (Future<Object?>.value(3) as FutureOr<Object));
```
which ends up not breaking soundness, with
```dart
Object o = await (Future<Object?>.value(null) as FutureOr<Object));
```
which currently does break soundness.

This change will make both into runtime errors.

(The alternative, and current specification,
to make both expression evaluate to the future itself,
is considered just as error prone and harder to debug,
since a type error is likely happen further downstream
when the `Object` is unexpectedly still a `Future`.)