Add canonicalization rules for non-records.
Currently a non-record is canonicalized based on the `identical` comparison of its fields against previously created constant values of the same type. 
Since a record field does not have a predictably or useful `identical` comparison result, we instead define a comparison structurally on the record, and use that instead.

Also formatting tweaks.