[patterns] Tease apart type testing from refutability.
Some patterns fail to match if the value doesn't have a certain type.
We still allow those patterns in irrefutable contexts if the matched
value is statically known to always have that type (or be dynamic).

Instead of specifying by having the refutability state of those patterns
depend on their type, make refutability a syntactic property and then
define the compile-time error for the types during type checking.

Fix #2450.