Move the feature specification of super parameters into `accepted` (#2514)
This PR copies the super parameters feature specification from 'working/.../proposal.md' to 'accepted/2.17/super-parameters/feature-specification.md', and it replaces 'proposal.md' by a small file that just contains a link to the new location. It also changes the title from 'Super-Initializer Parameters' to 'Super Parameters', because that's the established terminology today.