Removed the special exception for a view on a top type: It is now a proper subtype of Object? (#2559)
The feature specification about views previously had the rule that a view whose representation type is a top type is itself a top type. This was motivated by the subtype relationship between the representation type and the view type that we used to have.

For example, with `view class V(Object? o) {}`, we had `Object? <: V` so we were forced to make `V` a top type.

We do not have that subtype relationship any more, so we're free to make such a view type a proper subtype of `Object?`. This PR does that. This prevents unchecked/silent assignments of the form `V v = whateverYouWant;` (which means that `V` yields no protection at all against adopting this view by mistake).