Write separate more detailed spec for sealed types.
I forked this off from the larger type-modifiers strawman for a few
reasons:

*   That strawman also proposed disallowing mixing in classes which is
    a larger breaking change that will require a migration. We almost
    certainly can't get that done in time for patterns.

*   There is a lot more motivation around keyword choices in that doc
    which is useful to have tracked, but not super relevant for a
    feature spec, at least not at that level of detail.

*   The "closed" and "base" modifiers in that proposal are small
    features with no breakage, and can likely be done quickly. But
    they'll still probably be behind a separate experiment flag from
    "sealed", so I thought it made sense to propose those separately.

I also changed the keyword from "switch" to "sealed". While some liked
the former, my general impression is that most don't and it didn't
seem to "stick". When we talk about the feature, we all seem to call it
"sealed", which is probably telling us something.