Patterns: adjust nomenclature of variable pattern types.
Change the "variables" part of section "Type checking and pattern
required type" to consistently refer to the unpromoted type of the
variable as its static type rather than its declared type.  Rationale:
in the implementations, the term "declared type" is consistently used
to refer to a type that's explicitly specified by the user for the
variable pattern (and is `null` if the type is inferred).

This change is consistent with the following text from the "Pattern
variable sets" section:

> Each variable in the set has a unique name, a static type (the
> declared or inferred type, but not its promoted type), and whether
> it is final or not.